fix: clippy warnings
- fix clippy warnings

Tested:
- Yes
