--- src/Ppmd8Enc.rs
@@ -1,28 +1,0 @@
-/* Ppmd.h -- PPMD codec common code
-2021-04-13 : Igor Pavlov : Public domain
-This code is based on PPMd var.H (2001): Dmitry Shkarin : Public domain */
-/* Most compilers works OK here even without #pragma pack(push, 1), but some GCC compilers need it. */
-/* SEE-contexts for PPM-contexts with masked symbols */
-
-/*
-   PPMD code can write full CPpmd_State structure data to CPpmd*_Context
-      at (byte offset = 2) instead of some fields of original CPpmd*_Context structure.
-
-   If we use pointers to different types, but that point to shared
-   memory space, we can have aliasing problem (strict aliasing).
-
-   XLC compiler in -O2 mode can change the order of memory write instructions
-   in relation to read instructions, if we have use pointers to different types.
-
-   To solve that aliasing problem we use combined CPpmd*_Context structure
-   with unions that contain the fields from both structures:
-   the original CPpmd*_Context and CPpmd_State.
-   So we can access the fields from both structures via one pointer,
-   and the compiler doesn't change the order of write instructions
-   in relation to read instructions.
-
-   If we don't use memory write instructions to shared memory in
-   some local code, and we use only reading instructions (read only),
-   then probably it's safe to use pointers to different types for reading.
-*/
-// PPMD_32BIT
